refactor: 직업 경험치를 100% 누적 방식으로 변경
- 모든 정상 직업 행동의 경험치 지급 확률을 100%로 통일
- 기존 육성 속도를 유지하도록 행동별 기대 경험치를 소수 단위로 환산
- 소수 경험치와 경험치 증가 배율을 누적하고 저장·동기화하도록 변경

+++ JobsPlusRemastered-1.21.10/common/src/main/java/com/daqem/jobsplus/client/gui/jobs/components/ExperienceScrollContentComponent.java
@@ -24,7 +24,7 @@ public class ExperienceScrollContentComponent extends EmptyComponent
 
         List<IAction> actions = state.getSelectedJob().getJobInstance().getActions().stream()
                 .filter(action -> action.getRewards().stream().anyMatch(JobExpReward.class::isInstance))
-                .sorted(Comparator.comparingInt(this::getMaximumExperience)
+                .sorted(Comparator.comparingDouble(this::getMaximumExperience)
                         .thenComparing(action -> action.getLocation().getPath()))
                 .toList();
 
@@ -59,12 +59,12 @@ public class ExperienceScrollContentComponent extends EmptyComponent
         this.setHeight(Math.max(0, yOffset - SECTION_GAP));
     }
 
-    private int getMaximumExperience(IAction action)
+    private double getMaximumExperience(IAction action)
     {
         return action.getRewards().stream()
                 .filter(JobExpReward.class::isInstance)
                 .map(JobExpReward.class::cast)
-                .mapToInt(JobExpReward::getMax)
+                .mapToDouble(JobExpReward::getMax)
                 .findFirst()
                 .orElse(0);
     }
